move MirrorPartitionState to server-common for both metadata and core used
